Redundante cast verwijderd.
git-svn-id: file:///home/kommusoft/tmpgit/swopsvn@416 c49496ef-df08-40b3-bdd7-5dbf4c3ba35c

+++ ProjectSWOP20102011/src/projectswop20102011/controllers/TimeAheadController.java
@@ -74,8 +74,8 @@ public class TimeAheadController extends Controller {
 		getWorld().getTimeSensitiveList().timeAhead(seconds);
 		if (getExternalSystem() != null) {
 			int s = (int) (seconds + getWorld().getTime());
-			int hours = (int) (s / 3600);
-			int minutes = (int) ((s - (3600 * hours)) / 60);
+			int hours = (s / 3600);
+			int minutes = ((s - (3600 * hours)) / 60);
 			ITime time = new Time(hours, minutes);
 			try {
 				getExternalSystem().notifyTimeChanged(time);
